+++ examples/customize_persistence_flow/src/aggregate.rs
@@ -5,7 +5,6 @@ use sqlx::{Pool, Postgres};
 
 use esrs::aggregate::{Aggregate, AggregateManager, AggregateState};
 use esrs::store::postgres::PgStore;
-use esrs::store::postgres::Projector;
 use esrs::store::StoreEvent;
 use esrs::types::SequenceNumber;
 
@@ -18,11 +17,12 @@ pub struct CounterAggregate {
 
 impl CounterAggregate {
     pub async fn new(pool: &Pool<Postgres>) -> Result<Self, CounterError> {
-        let projectors: Vec<Box<dyn Projector<Self> + Send + Sync>> = vec![Box::new(CounterProjector)];
+        let event_store: PgStore<CounterAggregate> = PgStore::new(pool.clone())
+            .setup()
+            .await?
+            .add_projector(Box::new(CounterProjector));
 
-        Ok(Self {
-            event_store: PgStore::new(pool.clone(), projectors, vec![]).setup().await?,
-        })
+        Ok(Self { event_store })
     }
 }