+++ examples/simple_side_effect/src/aggregate.rs
@@ -66,12 +66,13 @@ pub struct LoggingAggregate {
 
 impl LoggingAggregate {
     pub async fn new(pool: &Pool<Postgres>) -> Result<Self, LoggingError> {
-        let projectors: Vec<Box<dyn Projector<Self> + Send + Sync>> = vec![Box::new(LoggingProjector)];
-        let policies: Vec<Box<dyn Policy<Self> + Send + Sync>> = vec![Box::new(LoggingPolicy)];
+        let event_store: PgStore<LoggingAggregate> = PgStore::new(pool.clone())
+            .setup()
+            .await?
+            .add_projector(Box::new(LoggingProjector))
+            .add_policy(Box::new(LoggingPolicy));
 
-        Ok(Self {
-            event_store: PgStore::new(pool.clone(), projectors, policies).setup().await?,
-        })
+        Ok(Self { event_store })
     }
 }